refactor(admin): split src/lib/admin/server.ts by domain (CHAOS-1569)
